Remove sliding column content from main page to put it in navbar
